Move PartialRealizationConfig to comm package.
Moving PartialRealizationConfig from (private class in) dmod.dataservice
to dmod.communication package.
