+++ python/lib/communication/setup.py
@@ -21,6 +21,6 @@ setup(
     url='',
     license='',
     include_package_data=True,
-    install_requires=['dmod-core>=0.4.2', 'websockets>=8.1', 'jsonschema', 'redis', 'pydantic'],
+    install_requires=['dmod-core>=0.4.2', 'websockets>=8.1', 'jsonschema', 'redis', 'pydantic', 'ngen-config>=0.1.1'],
     packages=find_namespace_packages(include=['dmod.*'], exclude=['dmod.test'])
 )
