+++ python/services/dataservice/setup.py
@@ -17,8 +17,8 @@ setup(
     author_email='',
     url='',
     license='',
-    install_requires=['dmod-core>=0.5.0', 'dmod-communication>=0.13.0', 'dmod-scheduler>=0.11.0',
-                      'dmod-modeldata>=0.9.0', 'redis', "pydantic", "fastapi", "uvicorn[standard]",
-                      'ngen-config>=0.1.1'],
+    install_requires=["dmod-core>=0.5.0", "dmod-communication>=0.13.0", "dmod-scheduler>=0.11.0",
+                      "dmod-modeldata>=0.9.0", 'redis', "pydantic", "fastapi", "uvicorn[standard]",
+                      "ngen-config>=0.1.1"],
     packages=find_namespace_packages(exclude=['dmod.test', 'deprecated', 'conf', 'schemas', 'ssl', 'src'])
 )
